Fix StreamField inside InlinePanel
If a StreamField is used inside an InlinePanel, there will be no 'h2'
header anymore and we get js errors and things stop working.
This modification checks for the existance of the 'h2' tag and if not
found it will ignore this streamfield for creating useless (in this
case) language copy buttons.