Adds sub-module makemigrations to override django's 'makemigrations' command
New 'makemigrations' is a proxy for 'makemigrations_translation', it's
role is only to override django's own command and skip any
wagtailcore_page migrations caused by translation. Adding this to it's
own submodule makes this override optional for the user. The original
Django 'makemigrations' command is available by running
'makemigrations_original'